+++ C.novaextension/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Version 1.2.1
+
+This extension is superseded by **C-Dragon**.
+At start up we check to see if it is installed, provide
+guidance in switching to it.
+
 ## Version 1.2
 
 Basic indentation on block boundaries.